Extract shared types about relaying
