+++ packages/relay-kit/src/packs/gelato/GelatoRelayAdapter.ts
@@ -8,9 +8,9 @@ import {
   TransactionStatusResponse
 } from '@gelatonetwork/relay-sdk'
 import Safe from '@safe-global/safe-core-sdk'
-import { MetaTransactionData, SafeTransaction } from '@safe-global/safe-core-sdk-types'
+import { MetaTransactionData, MetaTransactionOptions, RelayTransaction, SafeTransaction } from '@safe-global/safe-core-sdk-types'
 import { GELATO_FEE_COLLECTOR, GELATO_NATIVE_TOKEN_ADDRESS, ZERO_ADDRESS } from '../../constants'
-import { MetaTransactionOptions, RelayAdapter, RelayTransaction } from '../../types'
+import { RelayAdapter } from '../../types'
 
 export class GelatoRelayAdapter implements RelayAdapter {
   #gelatoRelay: GelatoNetworkRelay
